A14: migrate polynomial coefficients from double to int.
All coefficients in PSMM are mathematically integers (typically drawn
from {-1, 0, 1}). The codebase historically represented them as
`std::vector<double>`, which cost a sprinkle of int<->double casts on
every path and hid two classes of subtle bug:

 - m_CoeffsIndices in the iterator was `std::map<int,int>` keyed by
   the truncated integer value of the (double) coefficient — any
   non-integer coefficient would silently collapse keys (A14 in the
   review);

 - H(p) / L(p) accumulation in polyproperties went through
   std::size_t(abs(a[k])) with a double `a[k]`, where the unqualified
   `abs` resolves to int abs(int) via <cstdlib> and would truncate
   non-integer magnitudes to zero (A15 in the review).

Fix by typing the abstraction honestly: coefficients are `int`
throughout — struct member, iterator alphabet + output, rootfinder
signatures, factorizer, property computations, loaders, tests.

Concretely:
  - reciprocal_polynomial_t::coeffs: vector<double> -> vector<int>.
  - reciprocal_polynomials_iterator's alphabet and emission buffer
    become vector<int>; the coefficient-index map stays
    map<int,int> (now always well-keyed).
  - compute_mahler_general_polynomial_d /
    compute_mahler_reciprocal_polynomial_d /
    estimate_mahler_reciprocal_polynomial_d: accept vector<int>.
  - mpsolve_compute_mahler_with_properties: accept `const int*`.
    Switch the MPSolve call from mps_monomial_poly_set_coefficient_d
    to _set_coefficient_int throughout -- removes the last
    inconsistency between rootfinder.h (was _d) and mahlerestimator.h
    (was _int).
  - factor_polynomial / factor_reciprocal_polynomial: accept / emit
    vector<int>; NTL's to_long already returns long, so the cast from
    double is gone.
  - is_primitive_polynomial / is_primitive_reciprocal_polynomial:
    both `coeffs` and `divisors` become vector<int>.
  - polyproperties compute_all_properties: NNZ is now counted
    directly over p.coeffs[1..N/2] (was std::accumulate with double
    accumulator); H/L loop uses std::abs(int) explicitly.
  - psmm.cpp: coefficient alphabet parsed via std::stoi (was std::stod);
    candidate/factor buffers become vector<int>; log/print paths drop
    the historical `int(poly[j])` casts.
  - Test helpers lose half_to_double(); MPSolve paths consume int
    coefficients directly.

Verification:
  - All 4 test binaries (310 assertions) green.
  - Integration smoke run
      `./psmm -degree=10 -coeffs=-1,1 -nnz=1,2,3,4 -threshold=1.3`
    finds the 8 expected entries, including Lehmer (1.1762808...).
  - Build is still warning-free under -Wall -Wextra.

No file-format change: both AllKnownParallel and Known180 write
integer coefficients anyway; readers now parse directly into int
instead of double -> int.

+++ mahlerestimator.h
@@ -11,7 +11,7 @@
 #ifndef PSMM_MAHLER_ESTIMATOR_H
 #define PSMM_MAHLER_ESTIMATOR_H
 
-inline double estimate_mahler_reciprocal_polynomial_d(const std::vector<double>& coeffs, double threshold, int nthreads = 1)
+inline double estimate_mahler_reciprocal_polynomial_d(const std::vector<int>& coeffs, double threshold, int nthreads = 1)
 {
     //
     // The function estimates Mahler measure of the reciprocal polynomial in double precision.
